Add interface for APIBaseClient
Problem:
There are three clients based off the APIBaseClient so three code paths
are needed to make the same type of call in each client

Solution:
Add an interface so any client that satisfies it can be used to make
these calls
